Atomic Services/pending-orders/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
supabase
redis
